requirements.txt
jieba==0.39
django==2.1.3
djangorestframework==3.9.0
django-taggit==0.23.0